+++ gerenciamentoEstudantes/src/main/java/tech/ada/java/gerenciamento/estudantes/gerenciamentoestudantes/Model/Estudante.java
@@ -48,7 +48,6 @@ public class Estudante {
     
     public Estudante(Boolean estaAtivo) {
         this.estaAtivo = true;
-        System.out.println("usou este esta ativo");
     }
 
     //Construtor - Editando dados cadastrados // PUT
@@ -61,6 +60,6 @@ public class Estudante {
         this.contatoResponsavel = contatoResponsavel;
         this.dataDeCadastro = LocalDateTime.now(ZoneId.of("UTC"));
         this.turma = turma;
-        System.out.println("usou este este aqui");
+    
     }
 }
